+++ src/modules/dates/aplication/create-day.use-case.spec.ts
@@ -1,14 +1,15 @@
 import { Test } from "@nestjs/testing";
-import { GenerateDatetime } from "./generate-datetime";
+import { CreateDayUseCase } from "./create-day.use-case";
 import { dateTimeMock, dateTimeWithBookingsMock } from "../test/mocks/date.mock";
+import { GenerateDatetime } from "../dateTime-build/generate-datetime";
 
-describe('GIVEN generate-datetime', () => {
-    let generateDatetime: GenerateDatetime;
-    beforeEach(async () => {
+describe('GIVEN create-day.use-case', () => {
+    let createDayUseCase: CreateDayUseCase;
+    beforeEach(async() => {
         const module = await Test.createTestingModule({
             providers: [
                 {
-                    provide: GenerateDatetime,
+                    provide:CreateDayUseCase,
                     useValue: {
                         createDateTime: jest.fn(() => dateTimeMock),
                         createOneDayWithBookings: jest.fn(() => dateTimeWithBookingsMock),
@@ -16,40 +17,34 @@ describe('GIVEN generate-datetime', () => {
                         createOneDay: jest.fn(() => 'sábado 26 de julio 2025 26/07/2025'),
                     },
                 },
+                GenerateDatetime
             ],
         }).compile();
 
-        generateDatetime = module.get<GenerateDatetime>(GenerateDatetime);
-    });
+        createDayUseCase = module.get<CreateDayUseCase>(CreateDayUseCase);
+    })
 
     it('SHOULD be defined', () => {
-        expect(generateDatetime).toBeDefined();
-    });
+        expect(createDayUseCase).toBeDefined();
+    })
 
     describe('WHEN createDateTime is called', () => {
         it('SHOULD return the matrix of datetime', () => {
-            const result = generateDatetime.createDateTime();
+            const result = createDayUseCase.createDateTime();
             expect(result).toEqual(dateTimeMock);
         });
     })
 
     describe('WHEN createOneDayWithBookings is called', () => {
         it('SHOULD return the matrix of datetime with bookings', () => {
-            const result = generateDatetime.createOneDayWithBookings();
+            const result = createDayUseCase.createOneDayWithBookings();
             expect(result).toEqual(dateTimeWithBookingsMock);
-        });
-    })
-
-    describe('WHEN createOneDay is called', () => {
-        it('SHOULD return the day', () => {
-            const result = generateDatetime.createOneDay();
-            expect(result).toEqual('sábado 26 de julio 2025 26/07/2025');
         });
     })
 
     describe('WHEN createNextDay is called', () => {
         it('SHOULD return the next day', () => {
-            const result = generateDatetime.createNextDay(new Date('2025-07-29T03:00:00.000Z'));
+            const result = createDayUseCase.createNextDay(new Date('2025-07-29T03:00:00.000Z'));
             expect(result).toEqual('miércoles 30 de julio 2025 30/07/2025');
         });
     })