+++ src/modules/dates/service/dates.service.ts
@@ -3,6 +3,7 @@ import { GoogleSheetsService } from 'src/shared/service/google-sheets.service';
 import { CreateDayUseCase } from '../aplication/create-day.use-case';
 import { SHEETS_NAMES } from 'src/constants/sheets-name/sheets-name';
 import { parseDate } from '../utils/parseDate';
+import { DateTime } from 'src/lib/datetime/datetime.type';
 
 @Injectable()
 export class DatesService {
@@ -10,11 +11,11 @@ export class DatesService {
   constructor(
     private readonly createDayUseCase: CreateDayUseCase,
     private readonly googleSheetsService: GoogleSheetsService,
-  ) {}
-  async createDate():Promise<string> {
-    try {      
+  ) { }
+  async createDate(): Promise<string> {
+    try {
       const dateTime = this.createDayUseCase.createDateTime();
-      
+
       const dateTimeWithBookings = this.createDayUseCase.createOneDayWithBookings();
 
       await this.googleSheetsService.appendRow(`${SHEETS_NAMES[0]}!A:C`, dateTime);
@@ -29,18 +30,18 @@ export class DatesService {
     }
   }
 
-  async createNextDate():Promise<string> {
+  async createNextDate(): Promise<string> {
     try {
-      const lastRow = await this.googleSheetsService.getLasRowValue(`${SHEETS_NAMES[0]}!A:A`);      
+      const lastRow = await this.googleSheetsService.getLasRowValue(`${SHEETS_NAMES[0]}!A:A`);
       
       const parsedDate = parseDate(lastRow);
-      
-      const nextDay = this.createDayUseCase.createNextDay(parsedDate);
-      
-      const dateTime = this.createDayUseCase.createDateTime(nextDay);
 
-      const nextDayWithBookings = this.createDayUseCase.createOneDayWithBookings(nextDay);
-      
+      const nextDay = this.createDayUseCase.createNextDay(parsedDate);
+
+      const dateTime: DateTime = this.createDayUseCase.createDateTime(nextDay);
+
+      const nextDayWithBookings: DateTime = this.createDayUseCase.createOneDayWithBookings(nextDay);
+
       await this.googleSheetsService.appendRow(`${SHEETS_NAMES[0]}!A:C`, dateTime);
 
       await this.googleSheetsService.appendRow(`${SHEETS_NAMES[1]}!A:E`, nextDayWithBookings);
@@ -53,7 +54,29 @@ export class DatesService {
     }
   }
 
-  removeDate(id: number) {
-    return `This action removes a #${id} date`;
+  async createXDates(quantity: number): Promise<string> {
+    try {
+      for (let i = 0; i < quantity; i++) {
+        const lastRow = await this.googleSheetsService.getLasRowValue(`${SHEETS_NAMES[0]}!A:A`);
+
+        const parsedDate = parseDate(lastRow);
+
+        const nextDay = this.createDayUseCase.createNextDay(parsedDate);
+
+        const dateTime: DateTime = this.createDayUseCase.createDateTime(nextDay);
+
+        const nextDayWithBookings: DateTime = this.createDayUseCase.createOneDayWithBookings(nextDay);
+
+        await this.googleSheetsService.appendRow(`${SHEETS_NAMES[0]}!A:C`, dateTime);
+
+        await this.googleSheetsService.appendRow(`${SHEETS_NAMES[1]}!A:E`, nextDayWithBookings);
+
+      }
+      this.logger.log(`Se agregaron ${quantity} dias`, DatesService.name);
+      return `Se agregaron ${quantity} dias`;
+    } catch (error) {
+      this.logger.error(`Error al agregar el dia`, error);
+      throw error;
+    }
   }
 }