+++ src/modules/dates/test/mocks/date.mock.ts
@@ -1,4 +1,6 @@
-export const dateTimeMock = [
+import { DateTime } from "src/lib/datetime/datetime.type"
+
+export const dateTimeMock: DateTime = [
     [ '', '' ],
     [ '', '' ],
     [ 'sábado 26 de julio 2025 26/07/2025', '12:00' ],
@@ -14,3 +16,21 @@ export const dateTimeMock = [
     [ 'sábado 26 de julio 2025 26/07/2025', '22:00' ],
     [ '', '' ]
   ]
+
+  export const dateTimeWithBookingsMock: DateTime = [
+    [ '', '' ],
+    [ '', '' ],
+    [ 'sábado 26 de julio 2025 26/07/2025', '12:00', '0', '20' ],
+    [ 'sábado 26 de julio 2025 26/07/2025', '13:00', '0', '20' ],
+    [ 'sábado 26 de julio 2025 26/07/2025', '14:00', '0', '20' ],
+    [ 'sábado 26 de julio 2025 26/07/2025', '15:00', '0', '20' ],
+    [ 'sábado 26 de julio 2025 26/07/2025', '16:00', '0', '20' ],
+    [ 'sábado 26 de julio 2025 26/07/2025', '17:00', '0', '20' ],
+    [ 'sábado 26 de julio 2025 26/07/2025', '18:00', '0', '20' ],
+    [ 'sábado 26 de julio 2025 26/07/2025', '19:00', '0', '20' ],
+    [ 'sábado 26 de julio 2025 26/07/2025', '20:00', '0', '20' ],
+    [ 'sábado 26 de julio 2025 26/07/2025', '21:00', '0', '20' ],
+    [ 'sábado 26 de julio 2025 26/07/2025', '22:00', '0', '20' ],
+    [ '', '' ]
+  ]
+